+++ test/sequelize.spec.js
@@ -5,7 +5,16 @@ var bluebird = require('bluebird');
 var proxyquire = require('proxyquire').noCallThru();
 
 var ModelMock = function () {};
-var UtilsMock = {};
+var PathMock = {
+	normalize: function (p) { return p },
+	resolve: function (p) { return p },
+	dirname: function (p) { return p },
+};
+
+var UtilsMock = {
+	stack: function () { return {}; },
+};
+
 var PackageMock = {
 	version: 'test',
 };
@@ -15,13 +24,22 @@ var ErrorMock = {
 };
 var QueryInterfaceMock = function () {};
 
+var lastImportTestCall;
+function importTestFunc() {
+	lastImportTestCall = arguments;
+}
+
 var Sequelize = proxyquire('../src/sequelize', {
+	'path'     : PathMock,
 	'./model'  : ModelMock,
 	'./utils'  : UtilsMock,
 	'./errors' : ErrorMock,
 	'./queryinterface' : QueryInterfaceMock,
 	'../package.json'  : PackageMock,
-	'./data-types'     : function () {}
+	'./data-types'     : function () {},
+	
+	'import-test'      : importTestFunc,
+	'import-test-es6'  : { default: importTestFunc },
 });
 
 describe('Sequelize', function () {
@@ -141,7 +159,7 @@ describe('Sequelize', function () {
 			seq = new Sequelize();
 		});
 		
-		it('should queue a result against the QueryInterface', function () {
+		it('should clear queue of results in the QueryInterface', function () {
 			var run = 0;
 			seq.queryInterface = {
 				$clearQueue: function (res) {
@@ -150,6 +168,19 @@ describe('Sequelize', function () {
 			};
 			seq.$clearQueue();
 			run.should.equal(1);
+		});
+	});
+	
+	describe('#$overrideImport', function () {
+		var seq;
+		beforeEach(function () {
+			seq = new Sequelize();
+		});
+		
+		it('should override an import path in the importCache', function () {
+			seq.importCache = {};
+			seq.$overrideImport('foo', 'bar');
+			seq.importCache.should.have.property('foo').which.is.exactly('bar');
 		});
 	});
 	
@@ -190,7 +221,72 @@ describe('Sequelize', function () {
 			seq.isDefined('test').should.be.false()
 		});
 	});
+	
+	describe('#import', function () {
+		var seq, resolve, stack;
+		beforeEach(function () {
+			seq = new Sequelize();
+			lastImportTestCall = null;
+			resolve = PathMock.resolve;
+			stack = UtilsMock.stack;
+		});
+		
+		afterEach(function () {
+			PathMock.resolve = resolve;
+			UtilsMock.stack = stack;
+		});
+		
+		it('should return an already imported model', function () {
+			var findItem = {};
+			seq.importCache = {
+				'foo': findItem,
+			};
+			seq.import('foo').should.be.exactly(findItem);
+		});
+		
+		it('should import a model from the given path', function () {
+			seq.import('import-test');
+			should(lastImportTestCall).not.be.Null();
+			lastImportTestCall[0].should.be.exactly(seq);
+		});
+		
+		it('should turn a relative path into an absolute path', function () {
+			var pathRun = 0;
+			var stackRun = 0;
+			PathMock.resolve = function () { pathRun++; return './bar'; };
+			UtilsMock.stack = function () { stackRun++; return [0, { getFileName: function () { return 'baz'; } } ] };
+			var findItem = {};
+			
+			seq.importCache = {
+				'./bar': findItem,
+			};
+			seq.import('./foo').should.be.exactly(findItem);
+			pathRun.should.be.exactly(2);
+			stackRun.should.be.exactly(1);
+		});
+		
+		it('should import a replaced model from an overridden import', function () {
+			var findItem = {};
+			seq.importCache = {
+				'foo': 'bar',
+				'bar': findItem,
+			};
+			seq.import('foo').should.be.exactly(findItem);
+		});
 
+		it('should import an es6 model from the given path', function () {
+			seq.import('import-test-es6');
+			should(lastImportTestCall).not.be.Null();
+			lastImportTestCall[0].should.be.exactly(seq);
+		});
+
+		it('should import a model function as the second argument (for meteor compatibility)', function () {
+			seq.import('import-test', importTestFunc);
+			should(lastImportTestCall).not.be.Null();
+			lastImportTestCall[0].should.be.exactly(seq);
+		});
+	});
+	
 	describe('#model', function() {
 		it('should return a previously defined Mock Model referenced its name', function() {
 			var seq = new Sequelize();
@@ -257,4 +353,10 @@ describe('Sequelize', function () {
 		});
 	});
 	
+	describe('#authenticate', function () {
+		it('should simply return a resolving promise', (done) => {
+			var seq = new Sequelize();
+			seq.authenticate().then(done).catch(done);
+		})
+  });
 });